Create form.vb
create an application that uses the InputBox method to let the user enter 10 integers into an array. the application should display the largest and smallest values stored in the array. the following diagram shows a example of the application's form after all 10 values have been entered, wit the largest and smallest values displayed.